Change data type of Id in model & provided missing enums

+++ src/main/java/com/grad/sam/model/Account.java
@@ -21,7 +21,7 @@ public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "account_id")
-    private Long accountId;
+    private Integer accountId;
 
     @Column(name = "account_number", nullable = false, unique = true, length = 40)
     private String accountNumber;
